Consolidate the four scattered validators into a shared TestBuildValidations module
validate_mocks_in_use/validate_partials_in_use (TestBuildSetup),
validate_header_includes (TestBuildPlanner), and validate!
(TestSourceFileDirectiveResolver) were each independently defined,
identical in shape (raise CeedlingException on a real, user-facing
configuration mistake) but with no shared home -- same pattern
TestPipelineHelpers already covers for non-validation helpers.
