+++ lib/ceedling/test_invoker/test_source_file_directive_resolver.rb
@@ -8,6 +8,7 @@
 require 'ceedling/constants'
 require 'ceedling/exceptions'
 require 'ceedling/file_path_utils'
+require 'ceedling/test_invoker/test_build_validations'
 
 # Owns everything about the TEST_SOURCE_FILE() build directive's own path
 # content: collecting a test's raw directive strings, telling additive
@@ -17,6 +18,8 @@ require 'ceedling/file_path_utils'
 # valid; they don't need to know the +:/-: convention or how a raw string
 # becomes a real filepath.
 class TestSourceFileDirectiveResolver
+
+  include TestBuildValidations
 
   constructor :test_context_extractor, :file_finder, :configurator, :loginator
 
@@ -46,35 +49,6 @@ class TestSourceFileDirectiveResolver
     end
 
     return [additive_sources, subtractive_sources]
-  end
-
-  # A TEST_SOURCE_FILE() entry must name a real, recognized file whether it
-  # adds or removes -- a typo or a nonexistent path is worth failing loudly
-  # on either side of the +:/-: convention, the same way a missing file has
-  # always been treated for a plain, additive entry.
-  def validate!(test:, filepath:)
-    sources = @test_context_extractor.lookup_build_directive_sources_list( filepath )
-
-    ext_message = @configurator.extension_source.to_s
-    ext_message += " or #{@configurator.extension_assembly}" if @configurator.test_build_use_assembly
-
-    sources.each do |raw_source|
-      source = FilePathUtils.no_aggregation_decorators( raw_source )
-      valid_extension =
-        if @configurator.test_build_use_assembly
-          @configurator.extension_assembly.match?( source ) || @configurator.extension_source.match?( source )
-        else
-          @configurator.extension_source.match?( source )
-        end
-
-      unless valid_extension
-        raise CeedlingException.new( "File '#{source}' specified with TEST_SOURCE_FILE() in #{test} is not a #{ext_message} source file" )
-      end
-
-      if @file_finder.find_build_input_file( filepath: source, complain: :ignore, context: TEST_SYM ).nil?
-        raise CeedlingException.new( "File '#{source}' specified with TEST_SOURCE_FILE() in #{test} cannot be found in the source file collection" )
-      end
-    end
   end
 
   # Applies this test's TEST_SOURCE_FILE("-:...") entries against the fully
